Implement type/provider for vault_cert
This type/provider will obtain a certificate from vault on the client
side, using the host's Puppet certificate to authenticate to vault. The
Vault request and response are cached along with the destination paths
for the chain/cert/key in an info file. The info files live in a fixed
directory as determined by the `vault_cert_dir` fact. Use of info files
mean certs issed by vault_cert can be enumerated. This makes purging
obsolete resources possible.

The certificate expiration is cached as part of the vault issue
response, so renewal can be handled a configurable number of days
before a certificate expires, or if a certificate is found to already
be expired.

The `cert_data` is also cached, so that changes can be detected, and
a new certificate issued, for example if new aliases are wanted.

The provider will default to writing the chain/cert/key files out
to the `vault_cert_dir` directory, but this can be overridden and these
files placed at any destination path. Ownership and file permissions
can be customised. Corrective changes will be correctly applied to these
files. If the path is changed, the provider will make best efforts to
clean up the old files.