Fix expired certificate renewal
The expiration property should be read-only, but we use a hack to
allow it to trigger a refresh once the cert passes the renewal
threshold. Previously the implicit setter function would allow
puppet to override the value of the expiration property as read
from the system with the magic value `:auto`. This would prevent
the expiration check from functioning.

This change overrides the setter method so that the value of
expiration is never actually changed in the provider state. Since
the expiration is not a property we directly manage, it does not
matter that we are suppressing the change. It does mean the concrete
expiration timestamp of the certificate is always available, which
makes later checks if the certificate is still valid work and
return the correct result.